requirements.txt
Flask==3.0.3
pymongo==4.10.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.2
pandas==2.2.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
selenium==4.25.0
colorama==0.4.6
sentry-sdk==2.18.0
fastapi==0.115.0
pywebpush==2.0.3
lxml==5.3.0
discord.py==2.4.0